move error lines code server-side
